start_ap/requirements.txt
Flask==3.0.3
pyinstaller
ruamel.yaml==0.16.5